pyproject.toml
[tool.poetry]
name = "portal-atividades"
version = "0.1.0"
description = "Portal colaborativo de atividades escoteiras"
authors = ["Guionardo Furlan <guionardo@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "portal_atividades"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
django-multiselectfield = "^0.1.12"
urllib3 = "^1.26.14"
uplink = "^0.9.7"
martor = "^1.6.16"
django-environ = "^0.9.0"
django-bootstrap-v5 = "^1.0.11"
marshmallow = "^3.19.0"
django-extensions = "^3.2.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
